pyproject.toml
[project]
name = "strands-agentcore-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.8",
    "botocore>=1.39.8",
    "fastapi>=0.115.6",
    "mcp>=1.6.0",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "rx>=3.2.0",
    "tzdata>=2024.2",
    "uvicorn>=0.34.0",
    "strands-agents>=1.0.0",
    "strands-agents[otel]",
    "strands-agents-tools[mem0_memory]>=0.2.1",
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "playwright",
    "nest_asyncio"
]
